pyproject.toml
[tool.poetry]
name = "microservices"
version = "0.1.0"
description = ""
authors = ["Oleg Korshunov <oleg.korshunov3000@gmail.com>"]
readme = "README.md"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.autoflake]
check = true

[tool.pyright]
include = ["src"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.99.1"}
sqlalchemy = "^2.0.17"
asyncpg = "^0.27.0"
isort = "^5.12.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.3.0"
autoflake = "^2.2.0"
pyright = "^1.1.316"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
